Migrate fjs/effects/module.f.ts to .f.mjs, split type-level API into types.ts
The core effect system module: Operation, Effect, Pure, Cont, Do,
History, OperationMap, MatchResult, ToAsyncOperationMap, Param, Return,
F, Func move to types.ts unchanged (real TypeScript, variance
annotations intact); the runtime combinators (pure, step, mapStep,
historyStep, history, do_, foldStep, forEachStep, okStep, runPure,
match) become JSDoc-typed .mjs, using @template/@param on multi-step
curried functions where an inline @type can't expose its generics to
the function body.

Updates all 30+ dependents across the repo (cas, ci, cli, dev, djs,
emergent_testing, mcp, media/type, protocol/mcp, text/sgr, website,
and effects' own submodules) to import runtime values from .f.mjs and
types from types.ts separately. proof.f.ts stays TypeScript for now.
