+++ fjs/cas/evo/module.f.ts
@@ -42,7 +42,8 @@
  *
  * @module
  */
-import { pure, foldStep, type Effect, type Operation } from '../../effects/module.f.ts'
+import { pure, foldStep } from '../../effects/module.f.mjs'
+import type { Effect, Operation } from '../../effects/types.ts'
 import { eff } from '../../effects/eff/module.f.ts'
 import { create, read, write, type Key, type MemOp } from '../../effects/memory/module.f.ts'
 import { collectRead, type Cas } from '../module.f.ts'
