+++ fjs/cas/proof.f.ts
@@ -3,7 +3,8 @@ import { length, maxLength, msb, vec, vec8 } from '../types/bit_vec/module.f.mjs
 import { cBase32ToVec, vecToCBase32 } from '../basen/cbase32/module.f.mjs'
 import { computeSync, sha256 } from '../crypto/sha2/module.f.mjs'
 import { fileCas, casAddFile, collectRead, type FileCasOperation, casUpload } from './module.f.ts'
-import { match, pure, runPure, step, type Effect } from '../effects/module.f.ts'
+import { match, pure, runPure, step } from '../effects/module.f.mjs'
+import type { Effect } from '../effects/types.ts'
 import { mkdir, writeFile, rm, readFile, type ReadFile, type WriteFile, type Rm, type Mkdir, type IoResult, access } from '../effects/node/module.f.ts'
 import type { Ok } from '../types/result/types.ts'
 import { error, ok } from '../types/result/module.f.mjs'
