+++ fjs/dev/module.f.ts
@@ -16,7 +16,8 @@ import {
 import { cmp as strCmp } from '../types/string/module.f.mjs'
 import type { StringMap } from '../types/object/types.ts'
 import { unwrap } from '../types/result/module.f.mjs'
-import { pure, step, type Effect } from '../effects/module.f.ts'
+import { pure, step } from '../effects/module.f.mjs'
+import type { Effect } from '../effects/types.ts'
 import { join, relativize, toPosix } from '../path/module.f.mjs'
 import { assert, assertEq } from '../asserts/module.f.mjs'
 import { emptyState, virtual, type Dir } from '../effects/node/virtual/module.f.ts'
