+++ fjs/dev/update/proof.f.ts
@@ -8,7 +8,7 @@ import { utf8 } from '../../text/module.f.mjs'
 import { readUtf8File } from '../../effects/node/module.f.ts'
 import { defaultNodeProgramOptions, emptyState, virtual } from '../../effects/node/virtual/module.f.ts'
 import { main, syncMcp } from './module.f.ts'
-import { step } from '../../effects/module.f.ts'
+import { step } from '../../effects/module.f.mjs'
 
 const mcp = '{"servers":{}}' as const
 const initial = {
