+++ fjs/effects/mock/module.f.ts
@@ -3,7 +3,8 @@
  *
  * @module
  */
-import { match, type Effect, type Operation, type Pr } from "../module.f.ts"
+import { match } from "../module.f.mjs"
+import type { Effect, Operation, Pr } from "../types.ts"
 
 /**
  * A synchronous, state-threading operation map. An entry takes the command's
