+++ fjs/effects/todo/effect-list-fold.md
@@ -26,7 +26,7 @@ combinators speak the wrong one. Three consequences:
 in the store lives in memory at once, purely because `foldStep` cannot consume
 anything else.
 
-**The layering is inverted.** `fjs/effects/module.f.ts` — the core effect module
+**The layering is inverted.** `fjs/effects/module.f.mjs` — the core effect module
 — imports `fjs/types/list` (line 95) for `fold` and `List`, and *nothing else in
 that file uses either*. The two fold combinators are the module's only dependency
 on the strict list type.
@@ -44,7 +44,7 @@ and the eight importers currently alias around the clash — `elEmpty` in
 `fjs/cas`, `fjs/cas/evo`, `fjs/mcp`; `emptyList` in `fjs/media/type/proof`.
 The rename has value independent of the rest of this issue.
 
-**2. Move `foldStep` / `forEachStep`** out of `fjs/effects/module.f.ts` and into
+**2. Move `foldStep` / `forEachStep`** out of `fjs/effects/module.f.mjs` and into
 `fjs/effects/list/module.f.ts`, retyped over `EffectList`:
 
 ```ts
@@ -64,7 +64,7 @@ Keep the step-variant shape — effect first, one argument per line when the cal
 wraps. That is not a style preference: a step variant is this module's `do`
 notation, so the argument list is a statement list in execution order and the
 effect comes first because it happens first. The rationale currently lives on
-`foldStep`'s JSDoc and in the `fjs/effects/module.f.ts` header; carry both over.
+`foldStep`'s JSDoc and in the `fjs/effects/module.f.mjs` header; carry both over.
 
 **3. Add a strict-list converter.** All six current call sites hold a strict
 list, so the move needs a way in:
@@ -76,9 +76,9 @@ export const fromList = <O extends Operation, T>(items: List<T>): EffectList<O, 
 built from `empty` / `nonEmpty`. The four `pure(...)` call sites below become
 `fromList(...)`.
 
-**Layering.** `fjs/effects/list/module.f.ts` already imports `pure`, `Effect`
-and `Operation` from `../module.f.ts`; adding `step` is the same direction, so
-no cycle. After the move, `fjs/effects/module.f.ts` drops its
+**Layering.** `fjs/effects/list/module.f.ts` already imports `pure` from
+`../module.f.mjs` and `Effect`/`Operation` from `../types.ts`; adding `step` is
+the same direction, so no cycle. After the move, `fjs/effects/module.f.mjs` drops its
 `fjs/types/list` import entirely and the core effect module no longer depends on
 the strict list type at all.
 
@@ -139,7 +139,7 @@ follow-up in `fjs/cas` (see *Related*), not part of this issue.
 - [ ] Move `foldStep` / `forEachStep` there, retyped over `EffectList`, carrying
       their JSDoc and the step-variant rationale.
 - [ ] Remove the now-unused `fjs/types/list` import from
-      `fjs/effects/module.f.ts`.
+      `fjs/effects/module.f.mjs`.
 - [ ] Migrate the six call sites; the four strict ones go through `fromList`.
 - [ ] Create `fjs/effects/list/proof.f.ts` with full coverage.
 - [ ] Re-scope or close [fold-stream-combinator](./fold-stream-combinator.md).
@@ -154,5 +154,5 @@ follow-up in `fjs/cas` (see *Related*), not part of this issue.
   it is specified over `List<T>` and will want the same treatment.
 - [write-closed-helpers](../../cas/todo/write-closed-helpers.md) — already
   blocked by `fold-stream-combinator`, so transitively affected.
-- `fjs/effects/module.f.ts` header — the step-variant / `do`-notation rationale
+- `fjs/effects/module.f.mjs` header — the step-variant / `do`-notation rationale
   that fixes the argument order.
