+++ fjs/effects/todo/fold-stream-combinator.md
@@ -118,7 +118,7 @@ consumers first; the writers follow only if the shape stays clean.
   abstraction; `foldStream` is the consumer-side generalization.
 - [allreduce-combinator](./allreduce-combinator.md) — sibling combinator for
   parallel effects.
-- `okStep` (`fjs/effects/module.f.ts`) — the step-adapter helper shape; this
+- `okStep` (`fjs/effects/module.f.mjs`) — the step-adapter helper shape; this
   combinator's per-chunk step is a Kleisli function of the same shape.
 - [write-closed-helpers](../../cas/todo/write-closed-helpers.md) — hoists
   `fileCas.write`'s remaining nested helpers; its loop conversion depends on
