+++ fjs/mcp/evo/proof.f.ts
@@ -7,7 +7,7 @@ import { vecToCBase32 } from '../../basen/cbase32/module.f.mjs'
 import { initEvo, evo, type Evo } from '../../cas/evo/module.f.ts'
 import { evoToolRegistry } from './module.f.ts'
 import type { ToolEntry, ToolsCallResult } from '../../protocol/mcp/module.f.ts'
-import type { Operation } from '../../effects/module.f.ts'
+import type { Operation } from '../../effects/types.ts'
 import { parse as parseJson } from '../../media/json/module.f.ts'
 import { array, string as rttiString } from '../../types/rtti/module.f.mjs'
 import { parse as rttiParse } from '../../types/rtti/parse/module.f.mjs'
