+++ fjs/mcp/module.f.ts
@@ -26,7 +26,8 @@
  *
  * @module
  */
-import { step, type Effect } from '../effects/module.f.ts'
+import { step } from '../effects/module.f.mjs'
+import type { Effect } from '../effects/types.ts'
 import { create, type MemOp } from '../effects/memory/module.f.ts'
 import { type Read, type Write } from '../effects/node/module.f.ts'
 import { stdioTransport } from '../protocol/mcp/stdio/module.f.ts'
