+++ fjs/media/type/module.f.ts
@@ -36,7 +36,8 @@ import type { Vec } from '../../types/bit_vec/types.ts'
 import { msb, fromSentinel, length, u8List } from '../../types/bit_vec/module.f.mjs'
 import { iterable } from '../../types/list/module.f.mjs'
 import type { Nullable } from '../../types/nullable/types.ts'
-import { pure, step, type Effect, type Operation } from '../../effects/module.f.ts'
+import { pure, step } from '../../effects/module.f.mjs'
+import type { Effect, Operation } from '../../effects/types.ts'
 import type { List } from '../../effects/list/module.f.ts'
 import type { IoResult } from '../../effects/node/module.f.ts'
 import { ok, error } from '../../types/result/module.f.mjs'
