+++ fjs/protocol/mcp/module.f.ts
@@ -16,7 +16,8 @@
 import { boolean, string, option, array, record, or } from '../../types/rtti/module.f.mjs'
 import { unknown, type Unknown } from '../../media/json/module.f.ts'
 import type { Ts } from '../../types/rtti/ts/types.ts'
-import { pure, type Operation, type Effect, step } from '../../effects/module.f.ts'
+import { pure, step } from '../../effects/module.f.mjs'
+import type { Operation, Effect } from '../../effects/types.ts'
 import { read, write, type Key, type MemOp } from '../../effects/memory/module.f.ts'
 import {
     decodeRequest,
