+++ fjs/protocol/mcp/proof.f.ts
@@ -1,7 +1,7 @@
 import { assert, assertEq } from '../../asserts/module.f.mjs'
-import { pure, step, type Operation } from '../../effects/module.f.ts'
+import { pure, step } from '../../effects/module.f.mjs'
 import { eff } from '../../effects/eff/module.f.ts'
-import type { Effect } from '../../effects/module.f.ts'
+import type { Effect, Operation } from '../../effects/types.ts'
 import { run, type MemOperationMap } from '../../effects/mock/module.f.ts'
 import { asBase, asNominal, create, read, type Key, type MemOp } from '../../effects/memory/module.f.ts'
 import type { Unknown } from '../../media/json/module.f.ts'
