+++ fjs/protocol/mcp/stdio/module.f.ts
@@ -29,7 +29,8 @@
  *
  * @module
  */
-import { pure, step, type Effect, type Operation } from '../../../effects/module.f.ts'
+import { pure, step } from '../../../effects/module.f.mjs'
+import type { Effect, Operation } from '../../../effects/types.ts'
 import { readLine, write, type IoResult, type Read, type Write } from '../../../effects/node/module.f.ts'
 import { tryUtf8 } from '../../../text/module.f.mjs'
 import { parse, stringify, type Unknown } from '../../../media/json/module.f.ts'
