+++ fjs/website/module.f.ts
@@ -5,7 +5,8 @@
  */
 import { htmlUtf8 } from '../media/html/module.f.mjs'
 import { writeFile, type WriteFile } from '../effects/node/module.f.ts'
-import { pure, step, type Effect } from '../effects/module.f.ts'
+import { pure, step } from '../effects/module.f.mjs'
+import type { Effect } from '../effects/types.ts'
 import type { Vec } from '../types/bit_vec/types.ts'
 
 const html: Vec = htmlUtf8()(
